backend/requirements.txt
fastapi
uvicorn
python-dotenv
python-multipart
canvasapi
google-api-python-client
google-auth-oauthlib
google-genai
PyMuPDF
pydantic
pydantic-settings
supabase
pytest
httpxcryptography
